+++ src/client/Microsoft.Identity.Client.Services/Kerberos/KerberosSupplementalTicket.cs
