+++ src/client/Microsoft.Identity.Client.Services/Kerberos/KerberosTicketContainer.cs
